[FE feat: fe/refactortype

See merge request s10-final/S10P31A607!94

+++ frontend/abletotrip/app/src/main/java/com/hungrybrothers/abletotrip/ui/datatype/Catalog2Attractions.kt
@@ -13,7 +13,7 @@ data class Catalog2Attraction(
     val gu: String,
     val dong: String,
     val image_url: String,
-    val distance: Int,
+    val distance: Float,
 )
 
 @Serializable
